requirements.txt
python-telegram-bot==20.0
openai==0.27.0
configparser==5.3.0
transformers==4.31.0
torch==2.0.1
datasets==2.14.4
logging==0.5.1.2
